{{ cookiecutter.project_name }}/requirements.txt
psycopg2-binary
django
django-extensions
django-crispy-forms
django-widget-tweaks
ujson
pyyaml
pymongo
